Refactored logger macro into a static function. Added file logging but unfortunately it's crashing (DLL boundary most likely). Added factory for ZstCable to keep allocations inside core.dll. Fixed missing msgpack float types when unpacking ZstValue. Fixed some unneccassary allocations of ZstURI causing memory leaks. Fixed crash caused by client deleting cable too early by receiving stage update message before future had been resolved.

+++ src/dependencies/Queue.h
@@ -40,7 +40,7 @@ class Queue
     queue_.pop();
   }
 
-  int size() {
+  size_t size() {
 	  return queue_.size();
   }
